backend/requirements.txt
Flask==3.1.0
python-dotenv==1.0.0
flask-cors==4.0.0
pymongo==4.6.1
authlib
requests